api/requirements.txt

Flask
python-dotenv
pymongo
gunicorn
langchain-groq
langchain-huggingface
langchain-chroma
pypdf